Updated RootScanner.cpp
The RootScanner now updates the canvas during the idle task.
If an UpdateZoom was called with no zoomInfo a ResetZoom is called to
avoid setting the default zoom to 0,1. Changed determination of user zoom
from checking that the zoom values are not equal axis values to checking that
they are smaller than than the axis values. This resolves an issue where the
pad bounds may be extnded and updatezoom believes the user did this.