Changed HistScanner to store tree in arrays.
Changed to arrays as the vector have the downside that plotting on
versus another is difficult. Arrays are a bit more flexible, but have
the downside that clearing is more complex and that the tree needs to be
gated to avoid significant number of 0 counts for unfired channels.

Updated plot command to take module / channel number or not.