Adding new histograms. Updating master.xml
The master.xml has been updated to reflect the new status of the setup. 

Histograms have been added to allow for quicker online debugging of the PID plots. These plots include those gated with implants and those gated with the Ge. 

The E14060Processor now accepts a pair<double,double> providing a range of Ge energies to gate the PID plots on. These can also be used to potentially gate VANDLE spectra.